test(modals): complete migration to ports & adapters architecture
- Update 5 remaining legacy modal test files to use new architecture
- Remove global modal mocks from test/setup.ts
- Update camera-capture, media-studio, browser, media-restoration tests
- Clean up ModalProvider usage - no longer needed with new architecture
- Fix all test mocks to use useModals from @/domains/system-integration
- Update property names: modalType→activeModal, isOpen→isModalOpen

Legacy cleanup completed:
- src/features/camera-capture/__tests__/components/camera-capture-modal-screen.test.tsx
- src/features/media-studio/components/media-studio.test.tsx
- src/features/browser/__tests__/adapters/browser-adapter-mocks.ts
- src/test/setup.ts (removed 3 global modal mocks)
- src/features/media/__tests__/hooks/use-media-restoration.test.ts

All modal tests now use consistent architecture with useModals hook.

+++ src-tauri/src/state/commands/handler.rs
@@ -1740,8 +1740,8 @@ impl CommandHandler {
       bin: None,
       added_at: std::time::SystemTime::now()
         .duration_since(std::time::UNIX_EPOCH)
-        .map(|d| d.as_secs() as i64)
-        .unwrap_or(0),
+        .map(|d| d.as_secs() as f64)
+        .unwrap_or(0.0),
     };
 
     // Add to media pool
@@ -4880,8 +4880,8 @@ impl CommandHandler {
         bin: None,
         added_at: std::time::SystemTime::now()
           .duration_since(std::time::UNIX_EPOCH)
-          .map(|d| d.as_secs() as i64)
-          .unwrap_or(0),
+          .map(|d| d.as_secs() as f64)
+          .unwrap_or(0.0),
       };
 
       // Add to project if open
